+++ src/pygeoprocessing/geoprocessing_core.pyx
@@ -24,89 +24,85 @@ LOGGER = logging.getLogger('pygeoprocessing.geoprocessing_core')
 
 cdef float _NODATA = -1.0
 
-
-class MaskWrapper(object):
-    def __init__(self, base_nodata):
-        self.base_nodata = base_nodata
-
-    def __call__(self, base_array):
-        result = numpy.empty(base_array.shape, dtype=numpy.int8)
-        result[:] = _NODATA
-        valid_mask = base_array != self.base_nodata
-        result[valid_mask] = base_array[valid_mask] != 0
-        return result
-
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-def distance_transform_edt(base_mask_raster_path_band, target_distance_path):
-    """Calculate the Euclidean distance transform.
+def _distance_transform_edt(
+        region_raster_path, g_raster_path, double sample_d_x,
+        double sample_d_y, target_distance_raster_path):
+    """Calculate the euclidean distance transform on base raster.
+
+    Calculates the euclidean distance transform on the base raster in units of
+    pixels multiplied by an optional scalar constant. The implementation is
+    based off the algorithm described in:  Meijster, Arnold, Jos BTM Roerdink,
+    and Wim H. Hesselink. "A general algorithm for computing distance
+    transforms in linear time." Mathematical Morphology and its applications
+    to image and signal processing. Springer, Boston, MA, 2002. 331-340.
+
+    The base mask raster represents the area to distance transform from as
+    any pixel that is not 0 or nodata. It is computationally convenient to
+    calculate the distance transform on the entire raster irrespective of
+    nodata placement and thus produces a raster that will have distance
+    transform values even in pixels that are nodata in the base.
 
     Parameters:
-        base_mask_raster_path_band (tuple): a (path, band index) tuple to
-            calculate value from non-zero valued pixels.
-
-        target_distance_path (string): a path to a raster created by this
-        function with same dimensions and projection as base_mask_path where
-        all non-zero values of base_mask_path are equal to the euclidean
-        distance to the closest 0 pixel.
+        region_raster_path (string): path to a byte raster where region pixels
+            are indicated by a 1 and 0 otherwise.
+        g_raster_path (string): path to a raster created by this call that
+            is used as the intermediate "g" variable described in Meijster
+            et. al.
+        sampling_distance (float): parameter used to linearly scale the
+            pixel distances when calculating the distance transform.
+        target_distance_raster_path (string): path to the target raster
+            created by this call that is the exact euclidean distance
+            transform from any pixel in the base raster that is not nodata and
+            not 0. The units are in (pixel distance * sampling_distance).
 
     Returns:
-        None."""
+        None
+
+    """
     cdef int yoff, row_index, block_ysize, win_ysize, n_rows
     cdef int xoff, block_xsize, win_xsize, n_cols
     cdef int q_index, local_x_index, local_y_index, u_index
-    cdef int gu, gsq, tq, sq
-    cdef numpy.ndarray[numpy.int32_t, ndim=2] g_block
+    cdef int tq, sq
+    cdef float gu, gsq
+    cdef numpy.ndarray[numpy.float32_t, ndim=2] g_block
     cdef numpy.ndarray[numpy.int32_t, ndim=1] s_array
     cdef numpy.ndarray[numpy.int32_t, ndim=1] t_array
     cdef numpy.ndarray[numpy.float32_t, ndim=2] dt
     cdef numpy.ndarray[numpy.int8_t, ndim=2] mask_block
 
-    file_handle, base_mask_path = tempfile.mkstemp()
-    os.close(file_handle)
-    base_raster_info = pygeoprocessing.get_raster_info(
-        base_mask_raster_path_band[0])
-    base_nodata = base_raster_info['nodata'][
-        base_mask_raster_path_band[1]-1]
+    mask_raster = gdal.OpenEx(region_raster_path, gdal.OF_RASTER)
+    mask_band = mask_raster.GetRasterBand(1)
 
-    pygeoprocessing.raster_calculator(
-        [base_mask_raster_path_band], MaskWrapper(base_nodata),
-        base_mask_path,
-        gdal.GDT_Byte, _NODATA, calc_raster_stats=False)
+    n_cols = mask_raster.RasterXSize
+    n_rows = mask_raster.RasterYSize
 
-    base_mask_raster = gdal.OpenEx(base_mask_path)
-    base_mask_band = base_mask_raster.GetRasterBand(1)
-
-    n_cols = base_mask_raster.RasterXSize
-    n_rows = base_mask_raster.RasterYSize
-
-    file_handle, g_path = tempfile.mkstemp()
-    os.close(file_handle)
-    raster_info = pygeoprocessing.get_raster_info(
-        base_mask_raster_path_band[0])
+    raster_info = pygeoprocessing.get_raster_info(region_raster_path)
     pygeoprocessing.new_raster_from_base(
-        base_mask_raster_path_band[0], g_path, gdal.GDT_Int32, [_NODATA],
+        region_raster_path, g_raster_path, gdal.GDT_Float32, [_NODATA],
         fill_value_list=None)
-    g_raster = gdal.OpenEx(g_path, gdal.GA_Update)
+    g_raster = gdal.OpenEx(g_raster_path, gdal.OF_RASTER | gdal.GA_Update)
     g_band = g_raster.GetRasterBand(1)
     g_band_blocksize = g_band.GetBlockSize()
 
-    numerical_inf = (
+    # distances can't be larger than half the perimeter of the raster.
+    cdef float numerical_inf = sample_d_x * sample_d_y * (
         raster_info['raster_size'][0] + raster_info['raster_size'][1])
     # scan 1
     done = False
     block_xsize = raster_info['block_size'][0]
     mask_block = numpy.empty((n_rows, block_xsize), dtype=numpy.int8)
-    g_block = numpy.empty((n_rows, block_xsize), dtype=numpy.int32)
+    g_block = numpy.empty((n_rows, block_xsize), dtype=numpy.float32)
     for xoff in numpy.arange(0, n_cols, block_xsize):
         win_xsize = block_xsize
         if xoff + win_xsize > n_cols:
             win_xsize = n_cols - xoff
             mask_block = numpy.empty((n_rows, win_xsize), dtype=numpy.int8)
-            g_block = numpy.empty((n_rows, win_xsize), dtype=numpy.int32)
+            g_block = numpy.empty((n_rows, win_xsize), dtype=numpy.float32)
             done = True
-        base_mask_band.ReadAsArray(
+        mask_band.ReadAsArray(
             xoff=xoff, yoff=0, win_xsize=win_xsize, win_ysize=n_rows,
             buf_obj=mask_block)
         # base case
@@ -117,32 +113,34 @@ def distance_transform_edt(base_mask_raster_path_band, target_distance_path):
                     g_block[row_index, local_x_index] = 0
                 else:
                     g_block[row_index, local_x_index] = (
-                        g_block[row_index-1, local_x_index] + 1)
+                        g_block[row_index-1, local_x_index] + sample_d_y)
         for row_index in range(n_rows-2, -1, -1):
             for local_x_index in range(win_xsize):
                 if (g_block[row_index+1, local_x_index] <
                         g_block[row_index, local_x_index]):
                     g_block[row_index, local_x_index] = (
-                        1 + g_block[row_index+1, local_x_index])
+                        sample_d_y + g_block[row_index+1, local_x_index])
         g_band.WriteArray(g_block, xoff=xoff, yoff=0)
         if done:
             break
     g_band.FlushCache()
 
+    cdef float distance_nodata = -1.0
+
     pygeoprocessing.new_raster_from_base(
-        base_mask_raster_path_band[0], target_distance_path.encode('utf-8'),
-        gdal.GDT_Float32, [_NODATA], fill_value_list=None)
-    target_distance_raster = gdal.OpenEx(target_distance_path, gdal.GA_Update)
+        region_raster_path, target_distance_raster_path.encode('utf-8'),
+        gdal.GDT_Float32, [distance_nodata], fill_value_list=None)
+    target_distance_raster = gdal.OpenEx(
+        target_distance_raster_path, gdal.OF_RASTER | gdal.GA_Update)
     target_distance_band = target_distance_raster.GetRasterBand(1)
 
     LOGGER.info('Distance Transform Phase 2')
-
     s_array = numpy.empty(n_cols, dtype=numpy.int32)
     t_array = numpy.empty(n_cols, dtype=numpy.int32)
 
     done = False
     block_ysize = g_band_blocksize[1]
-    g_block = numpy.empty((block_ysize, n_cols), dtype=numpy.int32)
+    g_block = numpy.empty((block_ysize, n_cols), dtype=numpy.float32)
     dt = numpy.empty((block_ysize, n_cols), dtype=numpy.float32)
     sq = 0  # initialize so compiler doesn't complain
     gsq = 0
@@ -150,7 +148,7 @@ def distance_transform_edt(base_mask_raster_path_band, target_distance_path):
         win_ysize = block_ysize
         if yoff + win_ysize >= n_rows:
             win_ysize = n_rows - yoff
-            g_block = numpy.empty((win_ysize, n_cols), dtype=numpy.int32)
+            g_block = numpy.empty((win_ysize, n_cols), dtype=numpy.float32)
             dt = numpy.empty((win_ysize, n_cols), dtype=numpy.float32)
             done = True
         g_band.ReadAsArray(
@@ -166,7 +164,8 @@ def distance_transform_edt(base_mask_raster_path_band, target_distance_path):
                     tq = t_array[q_index]
                     sq = s_array[q_index]
                     gsq = g_block[local_y_index, sq]**2
-                    if ((tq-sq)**2 + gsq <= (tq-u_index)**2 + gu):
+                    if ((sample_d_x*(tq-sq))**2 + gsq <= (
+                            sample_d_x*(tq-u_index))**2 + gu):
                         break
                     q_index -= 1
                 if q_index < 0:
@@ -175,18 +174,19 @@ def distance_transform_edt(base_mask_raster_path_band, target_distance_path):
                     sq = u_index
                     gsq = g_block[local_y_index, sq]**2
                 else:
-                    w = 1 + (
-                        u_index**2 - sq**2 + gu - gsq) / (2*(u_index-sq))
-                    if w < n_cols:
+                    w = sample_d_x + (
+                        (sample_d_x*u_index)**2 - (sample_d_x*sq)**2 +
+                        gu - gsq) / (2*sample_d_x*(u_index-sq))
+                    if w < n_cols*sample_d_x:
                         q_index += 1
                         s_array[q_index] = u_index
-                        t_array[q_index] = w
+                        t_array[q_index] = <int>(w / sample_d_x)
 
             sq = s_array[q_index]
             gsq = g_block[local_y_index, sq]**2
             tq = t_array[q_index]
             for u_index in range(n_cols-1, -1, -1):
-                dt[local_y_index, u_index] = (u_index-sq)**2+gsq
+                dt[local_y_index, u_index] = (sample_d_x*(u_index-sq))**2+gsq
                 if u_index == tq:
                     q_index -= 1
                     if q_index >= 0:
@@ -204,15 +204,14 @@ def distance_transform_edt(base_mask_raster_path_band, target_distance_path):
         if done:
             break
 
+    target_distance_band.ComputeStatistics(0)
     target_distance_band.FlushCache()
-    gdal.Dataset.__swig_destroy__(target_distance_raster)
-    gdal.Dataset.__swig_destroy__(base_mask_raster)
-    gdal.Dataset.__swig_destroy__(g_raster)
-    try:
-        os.remove(g_path)
-    except OSError:
-        LOGGER.warning("couldn't remove file %s" % g_path)
-
+    target_distance_band = None
+    mask_band = None
+    g_band = None
+    target_distance_raster = None
+    mask_raster = None
+    g_raster = None
 
 @cython.boundscheck(False)
 @cython.wraparound(False)